Move back to subfolder
There is no way to set the include paths for use with the Halite compiler.
So, everything will be full relative paths for includes.
The compiler script Halite uses will climb down all the directories to
find the source files. It just does not set any include paths on the compile.